telegram-bot/requirements.txt
python-telegram-bot==21.6
google-cloud-aiplatform==1.78.0
google-cloud-firestore==2.19.0
firebase-admin==6.6.0
flask==3.1.0
gunicorn==23.0.0
